Moved registration functionality to own django app
