+++ malaria_api/tests.py
@@ -8,7 +8,7 @@ from rest_framework.renderers import JSONRenderer
 from rest_framework.test import APITestCase
 from malaria.models import Post
 from malaria_api.serializers import PostSerializer
-from webhub.models import Pcuser
+from signup.models import Pcuser
 
 
 class PostAPITestCase(APITestCase):
