+++ signup/templates/signup/footer.html
@@ -1,0 +1,65 @@
+<!-- 
+Version : Phython/Django 2.7.6, PostgreSQL 9.3.4
+Author : Vaibhavi Desai
+Github username : desaivaibhavi
+email : ranihaileydesai@gmail.com
+-->
+
+<!--the common footer of all tha pages has been formatted in this file -->
+
+<!--css portion -->
+
+<style>
+
+*{
+margin: 0;
+    }
+
+html, body {
+height: 100%;
+}
+
+.footerBkg {
+    background-color: #e2d7bb;
+    width: 100%;
+    padding: 30px 0px;
+    text-align: center;
+    font-size: 13px;
+    color: rgb(255, 255, 255);
+    line-height: 26px;
+}
+div{
+    margin: 0px;
+    padding: 0px;
+    border: 0px none;
+    font: inherit;
+    vertical-align: baseline;
+}
+
+</style>
+
+
+<!--html portion starts here -->
+<div class="footerBkg">
+   
+    
+			<div class="footerBtns"><font size="4" color="#2c0f00">
+				<a href="/aboutPC" title="Home" style="color:2c0f00;">About Peace Corps</a>&nbsp;&nbsp;|&nbsp;&nbsp;
+				<a href="/policies/" title="Know about our team" style="color:2c0f00;">Policies</a>&nbsp;&nbsp;|&nbsp;&nbsp;
+                <a href="/details/" title="FAQs" style="color:2c0f00;">Important Details</a>&nbsp;&nbsp;|&nbsp;&nbsp;
+				<a href="/helpPC/" title="Contact Us" style="color:2c0f00;">Help</a>&nbsp;&nbsp;
+				<br><br>
+				</font>
+            </div>
+    
+    
+
+    
+    
+    
+
+    
+			<p style="color:2c0f00;">&copy;2014 PeaceCrops</p>
+			
+			<a href="#top" style="color:2c0f00;"><font size="2">Return to Top</font></a>
+		</div>
