+++ signup/templates/signup/header.html
@@ -1,0 +1,29 @@
+<!-- 
+Version : Phython/Django 2.7.6, PostgreSQL 9.3.4
+Author : Vaibhavi Desai
+Github username : desaivaibhavi
+email : ranihaileydesai@gmail.com
+-->
+<style>
+    {% include "bootstrap/css/bootstrap.min.css" %}
+</style>
+<div class="navbar navbar-fixed-top" role="navigation" style="background-color:#e2d7bb; height:100px;">
+      <div class="container">
+        <div class="navbar-header">
+            <div class="row">
+                <div class="col-md-2">
+                    <a href="/"><img src="http://i.imgur.com/2wwH5x2.gif" class="img-responsive" style="width:180px; height:90px; padding-bottom:5px; padding-top:10px;"></a>
+                </div>
+                <div class="col-md-10">
+                    <a class="navbar-brand" href="/" style="color:#2c0f00;"><h1>Mobile App Control Center</h1></a>
+                    {% if pcuser %}
+                    <div class="navbar-right" style="float:right; padding-top:10px; margin-right:-500px;">
+                        <a href="/profile/?id={{pcuser.user.id}}"><h4><font color="#2c0f00" style="border:2px solid; padding:5px;">Welcome, <b>{{pcuser.user.first_name}}</b> !</font></a>
+                        <a href="/logout_do/"><font color="#2c0f00" style="border:2px solid; padding:5px;">LogOut</font></a></h4>
+                    </div>
+                    {% endif %}
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
