+++ webhub/admin.py
@@ -2,7 +2,6 @@ from django.contrib import admin
 from malaria.models import Post, RevPost
 from webhub.models import *
 
-admin.site.register(Pcuser)
 admin.site.register(Post)
 admin.site.register(RevPost)
